[refactor] Moved 'clearEventListeners' to its own TS file
