+++ src/Missions.js
@@ -3,7 +3,7 @@ import {Engine}                                     from "./engine";
 import {displayFactionContent}                      from "./Faction";
 import {Player}                                     from "./Player";
 import {dialogBoxCreate}                            from "../utils/DialogBox";
-import {clearEventListeners}                        from "../utils/HelperFunctions";
+import {clearEventListeners}                        from "../utils/uiHelpers/clearEventListeners";
 import {addOffset}                                  from "../utils/helpers/addOffset";
 import {formatNumber}                               from "../utils/StringHelperFunctions";
 import {getRandomInt}                               from "../utils/helpers/getRandomInt";
