+++ utils/FactionInvitationBox.js
@@ -1,7 +1,7 @@
 import {Faction, joinFaction}       from "../src/Faction";
 import {Engine}                     from "../src/engine";
 import {Player}                     from "../src/Player";
-import {clearEventListeners}        from "./HelperFunctions";
+import {clearEventListeners}        from "./uiHelpers/clearEventListeners";
 
 /* Faction Invitation Pop-up box */
 function factionInvitationBoxClose() {
